# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Social Vista API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/leads:
    get:
      operationId: listLeads
      tags: [admin]
      summary: List leads (admin)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: source
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [admin]
      summary: Create a lead (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /admin/leads/import:
    post:
      operationId: importLeads
      tags: [admin]
      summary: Bulk-import leads from a spreadsheet (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadImportResult"
  /admin/leads/{id}:
    patch:
      operationId: updateLead
      tags: [admin]
      summary: Update a lead (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [admin]
      summary: Delete a lead (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /admin/leads/{id}/activities:
    get:
      operationId: listLeadActivities
      tags: [admin]
      summary: List the activity timeline for a lead (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chronological activity log for the lead
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadActivity"
    post:
      operationId: createLeadActivity
      tags: [admin]
      summary: Add a manual activity log entry for a lead (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
  /admin/activities:
    get:
      operationId: listRecentActivities
      tags: [admin]
      summary: List the most recent activities across all leads (admin)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Most recent cross-lead activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentActivity"
  /admin/leads/{id}/reply:
    post:
      operationId: replyToLead
      tags: [admin]
      summary: Send an email reply to a lead (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadReplyInput"
      responses:
        "200":
          description: Reply sent; lead marked contacted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, status, source, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        serviceInterest:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        preferredTime:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        status:
          type: string
        source:
          type: string
        assignedTo:
          type: ["integer", "null"]
        lastContactedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serviceInterest:
          type: string
        message:
          type: string
        preferredTime:
          type: string
        source:
          type: string
    LeadImportInput:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/LeadInput"
        source:
          type: string
          description: Source label applied to every imported lead. Defaults to "import".
    LeadImportError:
      type: object
      required: [row, error]
      properties:
        row:
          type: integer
        error:
          type: string
    LeadImportResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/LeadImportError"
    LeadUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, booked, closed]
        assignedTo:
          type: ["integer", "null"]
        adminNotes:
          type: string
        markContacted:
          type: boolean
          description: When true, records the current time as the lead's last-contacted timestamp.
    LeadReplyInput:
      type: object
      required: [subject, message]
      properties:
        subject:
          type: string
          minLength: 1
          maxLength: 200
        message:
          type: string
          minLength: 1
          maxLength: 5000
        templateId:
          type: string
          enum: [intro, appointment, followup, proposal, promo]
    LeadActivity:
      type: object
      required: [id, leadId, type, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        type:
          type: string
        note:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
    RecentActivity:
      type: object
      required: [id, leadId, leadName, type, createdAt]
      properties:
        id:
          type: integer
        leadId:
          type: integer
        leadName:
          type: string
        type:
          type: string
        note:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadActivityInput:
      type: object
      required: [note]
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 5000
        type:
          type: string
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean