# lib/api-zod/src/generated/types/leadReplyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Social Vista API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadReplyInputTemplateId } from './leadReplyInputTemplateId';

export interface LeadReplyInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  subject: string;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  message: string;
  templateId?: LeadReplyInputTemplateId;
}